Adicionado a Junit

+++ src/calculadora/CalculadoraDivisao.java
@@ -1,7 +1,7 @@
 package calculadora;
 
 public class CalculadoraDivisao { 
-	public float dividir(float a, float b) {
+	public double dividir(double a, double b) {
 		if (a == 0 & b == 0) { // tratamento para quando receber dois valores Zeros
 			return 0;
 		} else if (a != 0 & b == 0) { // tratamento para quando receber o segundo numero Zero
